moves and renames in accordance to networking split.
