+++ org.palladiosimulator.analyzer.slingshot.stateexploration/src/org/palladiosimulator/analyzer/slingshot/stateexploration/explorer/networking/messages/SimTestMessage.java
@@ -1,0 +1,11 @@
+package org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.networking.messages;
+
+import org.palladiosimulator.analyzer.slingshot.networking.data.SimulationEventMessage;
+
+public class SimTestMessage extends SimulationEventMessage<String> {
+
+	public SimTestMessage(final String payload) {
+		super("TestSimEvent", payload);
+	}
+
+}
