+++ org.palladiosimulator.analyzer.slingshot.stateexploration/src/org/palladiosimulator/analyzer/slingshot/stateexploration/explorer/networking/messages/TestMessage.java
@@ -1,0 +1,13 @@
+package org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.networking.messages;
+
+import org.palladiosimulator.analyzer.slingshot.networking.data.EventMessage;
+
+/**
+ * Proof of Concept
+ */
+public class TestMessage extends EventMessage<String> {
+
+	public TestMessage(String payload) {
+		super("Test", payload);
+	}
+}
